% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{generate_formula}
\alias{generate_formula}
\title{Generate Standard Formulas}
\usage{
generate_formula(yvar, features, geo_features)
}
\arguments{
\item{yvar}{Target variable (e.g., "Pb_ppm")}

\item{features}{Vector of covariates (for ML/RK/Hier)}

\item{geo_features}{Vector of geo-self features (for FFRK)}
}
\value{
A list with \code{formula_main} and \code{formula_geo}
}
\description{
Generate Standard Formulas
}
