% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_base_methods_caret}
\alias{get_base_methods_caret}
\title{Get Base ML Methods Mapping for caret}
\usage{
get_base_methods_caret()
}
\value{
A named character vector mapping method aliases to caret method names.
}
\description{
Get Base ML Methods Mapping for caret
}
