% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_default_train_control}
\alias{get_default_train_control}
\title{Get Default caret::trainControl Object}
\usage{
get_default_train_control()
}
\value{
A caret trainControl object with default 5-fold CV and parallel support.
}
\description{
Get Default caret::trainControl Object
}
