% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{get_model_info}
\alias{get_model_info}
\title{Get Model Meta Information}
\usage{
get_model_info(model_type, formula_main = NULL, formula_geo = NULL)
}
\arguments{
\item{model_type}{A string (e.g., "rf", "FFRK_rf")}

\item{formula_main}{Formula for base models (e.g., Pb_ppm ~ x1 + x2 + ...)}

\item{formula_geo}{Formula for FFRK models (e.g., Pb_ppm ~ f.idw + f.sda.0 + ...)}
}
\value{
A list with type, caret method, and applicable formula
}
\description{
Returns modeling type and method name based on model identifier.
}
