% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_mae}
\alias{metric_mae}
\title{Calculate Mean Absolute Error (MAE)}
\usage{
metric_mae(observed, predicted)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
A numeric value representing MAE.
}
\description{
Computes the mean of absolute differences between observed and predicted values.
}
\examples{
metric_mae(c(1,2,3), c(1.1,1.9,3.2))

}
