% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_rmse}
\alias{metric_rmse}
\title{Calculate Root Mean Square Error (RMSE)}
\usage{
metric_rmse(observed, predicted)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
A numeric value representing RMSE.
}
\description{
Computes the root mean of squared differences between observed and predicted values.
}
\examples{
metric_rmse(c(1,2,3), c(1.1,1.9,3.2))

}
